+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 ### Patch Changes
 
 - Fixed: `module.(server.)ts` pairs not resolve at root src/lib
-  Fixed: `$rime/modules` types not generated on `rime generate` command
+- Fixed: `$rime/modules` types not generated on `rime generate` command